Using left click instead, signs on top
Using right clicks for interaction with fishy blocks didn't work very well when
adding placeable blocks, since they would appear in front of the fishy block
briefly, and it required that the player didn't stand right in front of the
block. Changed to left click.

We're now using wall posts on top of the block instead, which means that
signLocation no longer refers to the actual location of the sign, but rather
to the direction that the sign should face. Sign location is the block location
plus one on the y axis.