Added AutoSave Alert
It seems to be an ongoing (hourly) event where everyone spams chat with "LAGGG" "AUTOSAVE" bla bla. The code is like 3 lines so I thought it would be pointless in its own listener so I chucked it at the bottom of this. All it does is broadcast message when there is a save. Up to you whether you use but its there if you want.