BungeeCord plugin to link server chat, with optional IRC integration!
When I was talking to @EmilHernvall about BungeeCord a few months ago he said "we have to move all chat and message-related stuff to run on bungeecord". So, this is the first half of that, the "chat related stuff". Please be aware that this is _NOT_ a bukkit plugin like the commit i made for bungeecord was. This BungeeCord plugin goes in the plugins folder for the BungeeCord Server, for example while i put the last plugin in the directories "Bungee Server\Creative\plugins", "Bungee Server\Survival\plugins" and "Bungee Server\Hub\plugins". This plugin would go in the directory "Bungee Server\Bungee\plugins".

This commit includes:
* The base for the bungeecord plugin
* Cross Server chatting (can be turned on/off in settings.yml)
* IRC -> Server and Server -> IRC chat (can be turned on/off in settings.yml)
^^^ I only added this because Tregmine's own IRC is blocked at both school and work, so i cant stay connected :( ^^^
* ".online" IRC command (tells you player count and online players (i realise the online player list will prove an issue, if vanish isnt removed).
* ".version" IRC command (tells you the version of Minecraft that the network is running)
* A few other IRC related things like "ping" and "hello?"

Next commit (when I make a start on it):
* Cross server private messaging
* Cross server broadcasting?
* IRC -> Server. private messaging.

Here is a copy of the plugin setup on a demo network: https://docs.google.com/file/d/0B6zlQN0tXDtFZXYzV1dSdFZwTzg

To start the demo network up on your own machine you will need to open the ports (25564, 25565, 25566 and 25567) then run the 4 batch files that i made, one in each directory inside the "Bungee Server" folder.

I'll keep a demo network running for the next 12 hours or so. Here is the ip: 124.177.23.47
To switch between servers type "/server <server name>", and to view a list of servers on the demo network just type "/server".

To try out the cross server chat:
*  Log into the server with 2 different accounts, or get a friend to join with you
*  Make each go to a different server (maybe "/server creative" and "/server survival" or one of you can even stay in the hub server).
* Start chatting!

To try out the Server -> IRC and IRC -> Server chat:
*  Log into the server (you only need 1 account to try this)
*  Log into IRC and join #tregmine (host irc.esper.net, default port)
*  Type something in game and see it appear on IRC
*  Type something on IRC and see it appear ingame

To try out IRC commands:
*  Log into IRC
*  Type a command, ".version" for example (remember the dot before the command)
*  Try out an auto response "hello?" for example

NOTE: To build the plugin you will need to use Apache Maven, if you need any help building it, or you need help with any other part of the plugin then give me a buzz, As always I will be more than happy to help.